+++ tests/Enums/TestEnumInteger.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Pulsar\Tests\Enums;
+
+enum TestEnumInteger: int
+{
+    case First = 1;
+    case Second = 2;
+    case Third = 3;
+}