+++ tests/Enums/TestEnumString.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Pulsar\Tests\Enums;
+
+enum TestEnumString: string
+{
+    case First = 'first';
+    case Second = 'second';
+    case Third = 'third';
+}